+++ R/add_inline_forest_plot.R
@@ -3,7 +3,16 @@
 #' This function works with HTML output from the gt package only.
 #' Adds an in-line forest plot to a summary table.
 #'
+#' *Estimates from `tbl_regression()` and `tbl_uvregression()` that have
+#' been exponentiated are shown on the log scale.*
+#'
 #' @param x a gtsummary object
+#' @param header string indicating column header of new forest plot column.
+#' Default is `"**Forest Plot**"`.
+#' @param spec_pointrange.args named list of arguments that will be passed to
+#' `kableExtra::spec_pointrange()`. Use this argument to modify the
+#' default ascetics of the forest plot, e.g. color, size, symbols, etc.
+#' Default is `list(width = 250, cex = .75, col = "black", pch = 16)`
 #'
 #' @return gtsummary object
 #' @export
@@ -11,6 +20,7 @@
 #' @examples
 #' library(gtsummary)
 #'
+#' # Example 1 ----------------------------------
 #' add_inline_forest_plot_ex1 <-
 #'   lm(mpg ~ cyl + am + drat, mtcars) %>%
 #'   tbl_regression() %>%
@@ -19,40 +29,39 @@
 #' @section Example Output:
 #' \if{html}{\figure{add_inline_forest_plot_ex1.png}{options: width=80\%}}
 
-add_inline_forest_plot <- function(x) {
+add_inline_forest_plot <- function(x, header = "**Forest Plot**",
+                                   spec_pointrange.args = NULL) {
   # checking inputs ------------------------------------------------------------
-  if (!inherits(x, "gtsummary")) abort("`x=` must be class 'gtsummary'")
+  if (!inherits(x, "gtsummary")) rlang::abort("`x=` must be class 'gtsummary'")
   if (!all(c("estimate", "conf.low", "conf.high") %in% names(x$table_body)))
-    abort("`x$table_body` must contain columns 'estimate', 'conf.low', 'conf.high'")
+    rlang::abort("`x$table_body` must contain columns 'estimate', 'conf.low', 'conf.high'")
   assert_package("kableExtra", "add_inline_forest_plot()")
   updated_call_list <- c(x$call_list, list(add_inline_forest_plot = match.call()))
 
-  # set NULL value for plot ----------------------------------------------------
-  null_value <-
-    case_when(
-      inherits(x, c("tbl_regression", "tbl_uvregression")) ~
-        ifelse(x$inputs$exponentiate, 1L, 0L),
-      TRUE ~ 0L
-    )
+  # if exponentiated, plot on the log scale ------------------------------------
+  scale_fun <-
+    switch(
+      inherits(x, c("tbl_regression", "tbl_uvregression")) & x$inputs$exponentiate,
+      log
+    ) %||%
+    identity
 
   # add column with forest plot ------------------------------------------------
+  # prepping arguments for `kableExtra::spec_pointrange()`
+  spec_pointrange.args <-
+    list(vline = 0, width = 250, cex = .75, col = "black", pch = 16) %>%
+    purrr::list_modify(!!!spec_pointrange.args) %>%
+    purrr::list_modify(x = scale_fun(x$table_body$estimate),
+                       xmin = scale_fun(x$table_body$conf.low),
+                       xmax = scale_fun(x$table_body$conf.high))
+
   x <-
     gtsummary::modify_table_body(
       x = x,
       ~.x %>%
         dplyr::bind_cols(
           tibble::tibble(
-            forest_plot =
-              kableExtra::spec_pointrange(
-                x = .x$estimate,
-                xmin = .x$conf.low,
-                xmax = .x$conf.high,
-                vline = null_value,
-                width = 250,
-                cex = .75,
-                col = "black",
-                pch = 16
-              ) %>%
+            forest_plot = rlang::inject(kableExtra::spec_pointrange(!!!spec_pointrange.args)) %>%
               purrr::map("svg_text") %>%
               purrr::map(~gt::html(as.character(.x)))
           )
@@ -64,7 +73,7 @@ add_inline_forest_plot <- function(x) {
     gtsummary::modify_table_styling(
       columns = "forest_plot",
       hide = FALSE,
-      label = "**Forest Plot**"
+      label = header
     )
 
   # return updated gtsummary object --------------------------------------------